Put the CLI's Logfire exporter behind an off-by-default feature
`monty` is one binary serving two jobs: the developer CLI and the
`subprocess` worker the pool spawns. The worker never configures an
exporter — `main` dispatches to `subprocess::run` before telemetry setup,
deliberately, so an inherited LOGFIRE_TOKEN cannot reach a sandbox process
— but it still linked the whole OTLP stack: reqwest, rustls and aws-lc,
several MB of TLS the worker could not reach. Nothing tells the linker
which mode a binary will run in, so LTO cannot drop it.

`logfire` (and `tracing`, whose only use here is the span's deferred field)
is now optional behind `telemetry`, off by default. Enabled, behaviour is
exactly as before; disabled, LOGFIRE_TOKEN is ignored rather than
diagnosed, since a build without the feature has no opinion about a
variable set for other processes. `make lint-rs` already runs
`--all-features`, so the telemetry path stays checked.

The standalone path also moves to `run.rs`, leaving `main.rs` as the entry
point: allocator, style palette, clap definitions, and a `main` that parses
and branches. That makes the split structural — telemetry now lives in a
module the worker never enters.
