+++ crates/monty-runtime/Cargo.toml
@@ -16,6 +16,11 @@ repository = { workspace = true }
 name = "monty"
 path = "src/main.rs"
 
+[features]
+# CLI spans to Logfire. Off by default: its exporter links MBs of TLS that
+# `monty subprocess` — which never exports — would otherwise carry.
+telemetry = ["dep:logfire", "dep:tracing"]
+
 [dependencies]
 # Terminal styling for the CLI: `anstream` resolves NO_COLOR/CLICOLOR/tty and
 # enables virtual-terminal mode on Windows, `anstyle` builds the styles. Both
@@ -24,7 +29,7 @@ anstream = "0.6"
 anstyle = "1"
 clap = { workspace = true }
 # The standalone CLI owns process-level telemetry configuration.
-logfire = { version = "0.11", default-features = false, features = ["export-http-protobuf"] }
+logfire = { version = "0.11", default-features = false, features = ["export-http-protobuf"], optional = true }
 monty = { workspace = true }
 monty-fs = { workspace = true }
 monty-alloc = { workspace = true, features = ["exit-code"] }
@@ -32,7 +37,8 @@ monty-proto = { workspace = true, features = ["worker"] }
 monty-type-checking = { workspace = true }
 monty-types = { workspace = true }
 rustyline = "15"
-tracing = "0.1"
+# Only the `run monty` span's deferred field; goes with `logfire`.
+tracing = { version = "0.1", optional = true }
 
 [dev-dependencies]
 tempfile = "3"
